fix: lints and copyright checks

+++ pkg/pulumiyaml/syntax/encoding/object.go
@@ -232,7 +232,7 @@ func encodeValue(n syntax.Node, v reflect.Value) syntax.Diagnostics {
 
 		switch v.Kind() {
 		case reflect.Float32, reflect.Float64:
-			v.SetFloat(float64(n.Value()))
+			v.SetFloat(n.Value())
 		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
 			v.SetInt(int64(n.Value()))
 		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
